Expanded Icon Support, Changed Cheese.svg (#11)
* Rename gnome-settings.svg to org.gnome.tweaks.svg

* Delete org.Gnome.settings.svg

* Swapped old-settings icons with the symlinks to
preferences.svg

* Added Gnome Console (KGX) and BlackBox Icons

* Added Jellyfin.svg

* jellyfin flatpak symlink

* Symbolic link for text editor

* Matched colour pallete to the theme

The documents, and the pictures folder icons were off from the main pallet, fixed those

* Added Polari IRC icon

* Added Icon for Flowblade

cool history on this icon, it is actually borrowed from the actual project itself, so yes! thank the flowblade dev for me not having to trace the icon! ~~for sure not a flowblade simp~~

* updated Polari and Flowblade Icons

* Libreoffice, Firefox, and Evolution icons added, Cheese icon modified to reflect app colour